chore: move redux modules
