+++ src/entrypoints/main/redux/modules/catalog.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
-import { registrationApiGet } from '../../services/api/registration-api/host';
-import { reduxFsaThunk } from '../../lib/redux-fsa-thunk';
-import { catalogPath } from '../../services/api/registration-api/catalogs';
+import { registrationApiGet } from '../../../../services/api/registration-api/host';
+import { reduxFsaThunk } from '../../../../lib/redux-fsa-thunk';
+import { catalogPath } from '../../../../services/api/registration-api/catalogs';
 
 const CATALOG_REQUEST = 'CATALOG_REQUEST';
 export const CATALOG_SUCCESS = 'CATALOG_SUCCESS';
